requirements.txt
einops==0.8.0
ftfy==6.2.0
Pillow==10.0.1
regex==2023.10.3
scikit_learn==1.3.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
wandb==0.16.0
